Lexer Submodule

Lexer Submodule

SmartGameFormat.Lexer — Module.

The Lexer sub-module is concerned with transcribing a given stream of characters into a sequence of domain specific lexical units called "token".

Basic usage:

Wrap a plain IO object into a Lexer.TokenStream.

Call Lexer.next_token to collect another Lexer.Token.

Goto 2. unless end of file is reached.

source

Types

SmartGameFormat.Lexer.Token — Type.

Token(name::Char, [value::String])

A SGF specific lexical token. It can be either for the following:

Token('\0'): Empty token to denote trailing whitespaces.

Token(';'): Separator for nodes.

Token('(') and Token(')'): Delimiter for game trees.

Token('[') and Token(']'): Delimiter for property values.

Token('I', "AB1"): Identifier for properties. In general these are made up of one or more uppercase letters. However, with the exception of the first position, digits are also allowed to occur. This is done in order to supported older FF versions.

Token('S', "abc 23(\)"): Any property value between '[' and ']'. This includes moves, numbers, simple text, and text.

source

SmartGameFormat.Lexer.TokenStream — Type.

TokenStream(io::IO)

Stateful decorator around an io to create Token by using the function next_token.

source

Functions

SmartGameFormat.Lexer.next_token — Function.

next_token(ts::TokenStream) -> Token

Reads and returns the next Token from the given token stream ts. If no more token are available, then a EOFError will be thrown.

Note that the lexer should support FF[1]-FF[4] versions. In case any unambiguously illegal character sequence is encountered, the function will throw a LexicalError.

source

Exceptions

SmartGameFormat.Lexer.LexicalError — Type.

LexicalError(msg)

The string or stream passed to Lexer.next_token was not a valid sequence of characters according to the smart game format.

source