Parser Submodule

Parser Submodule

SmartGameFormat.Parser — Module.

The Parser sub-module is concerned with converting a sequence of Lexer.Token into a collection (i.e. a vector) of SGFGameTree.

To that end it provides the following functionality:

Parser.parse

Parser.tryparse

Parser.ParseError

source

Functions

SmartGameFormat.Parser.parse — Function.

parse(ts::Lexer.TokenStream) -> Vector{SGFGameTree}

Read all the lexial token from the stream ts into a Deque, and attempt to parse it as an SGF collection by calling tryparse. If successful, the SGF collection is returned as a vector of SGFGameTree. Note that the stream ts is first converted into a Deque{Token} in order to support peeking at the next Lexer.Token without removing it.

Depending on the content an exception may be thrown to signal that it is not a legal SGF specification.

Base.EOFError: Premature end-of-file encountered during tokenisation.

Lexer.LexicalError: illegal characters used outside property values. For example lower case letters for identifier.

Parser.ParseError: content is not a valid SGF specification (while considering the given the FF version).

source

SmartGameFormat.Parser.tryparse — Function.

tryparse(::Type{SGFNode}, seq::Deque{Token}) -> Nullable{SGFNode}

Try to parse the next N token in seq into a SGFNode, which means that the immediate next element in seq is expected to be Token(';') followed by zero or more properties. Each property must have a unique identifier, or a ParseError will be thrown.

source

tryparse(::Type{Pair}, seq::Deque{Token}) -> Nullable{Pair{Symbol,Vector{Any}}}

Try to parse the next N token in seq into a Pair denoting a single property of a SGFNode. Note that individual properties are parsed as Pair, because each SGFNode stores all its properties as a single Dict.

For a property to occur in seq, the immediate next element in seq must be a Token('I', "<ID>"), where <ID> is some sequence of uppercase letters denoting the identifier of the token. After the identifier there can be one or more property values. There must be at least one property value.

Each property value must be delimited by a Token('[') at the beginning and a Token(']') at the end. The value itself is contained within those two delimiter token as a single Token('S', "<val>") where <val> denotes the value. Note that this "S" token is optional and its absence means that the property value is the empty value.

source

tryparse(::Type{SGFGameTree}, seq::Deque{Token}) -> Nullable{SGFGameTree}

Try to parse the next N token in seq into a SGFGameTree.

A game tree must start with a Token('('), followed by one or more SGFNode, followed by zero or more sub-SGFGameTree, and finally end with a Token(')').

source

tryparse(::Type{Vector{SGFGameTree}}, seq::Deque{Token}) -> Nullable{Vector{SGFGameTree}}

Try to parse the next N token in seq as a Vector of SGFGameTree. Such a vector is called a "collection". For a collection to occur there must be at least one parse-able SGFGameTree in seq.

source

Exceptions

SmartGameFormat.Parser.ParseError — Type.

ParseError(msg)

The expression passed to Parser.parse could not be interpreted as a valid SGF specification (in accordance with the specified FF version).

source